Tidied up tab and added flyui.
Rejigged the tab using fields instead of labels to delineate the areas a little better. fix air (ex connect air, too long to fit on the tab, as well as 'release' ) and hand brake sprite moved to the top right of tab due to shifting tab names into sub title area. Everything appears to be working perfectly  but I didn't use the mods before so will need to be stress tested by someone that did,

This is the first time I have used C# (my last programming was jetDB for MS Access a quarter of a century ago!) and also the first time i've used github so apologies if I have stepped over a line by pushing this too far.